# app/api/team/route.js
export const dynamic = 'force-dynamic';
export const revalidate = 0;

import { NextResponse } from 'next/server';
import { getOrgId, getUserId } from '../../lib/org.js';
import { getSql } from '../../lib/db.js';
import { getTeamOrgAndMembers } from '../../lib/repositories/orgsTeam.repository.js';

// GET /api/team - List members + org info for caller's org
export async function GET(request) {
  try {
    const orgId = getOrgId(request);
    if (orgId === 'org_default') {
      return NextResponse.json(
        { error: 'Complete onboarding to manage your team', needsOnboarding: true },
        { status: 403 }
      );
    }

    const sql = getSql();

    const { org, members } = await getTeamOrgAndMembers(sql, orgId);
    if (!org) {
      return NextResponse.json({ error: 'Organization not found' }, { status: 404 });
    }

    const userId = getUserId(request);

    return NextResponse.json({
      org,
      members: members.map((m) => ({
        ...m,
        is_self: m.id === userId,
      })),
      member_count: members.length,
    });
  } catch (error) {
    console.error('Team API GET error:', error);
    return NextResponse.json({ error: 'Failed to fetch team' }, { status: 500 });
  }
}